% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame_func.R
\name{ampelgrafik_likert}
\alias{ampelgrafik_likert}
\title{Likert-style Traffic Light Chart with Percentages}
\usage{
ampelgrafik_likert(Data, Label)
}
\arguments{
\item{Data}{A numeric vector containing the data to be used for plotting.}

\item{Label}{The label for the x-axis.}
}
\value{
A likert-style traffic light chart with percentages showing the distribution of the data.
}
\description{
This function creates a likert-style traffic light chart with percentages from a given data vector and label.
The function first converts the data to numeric, calculates percentages, and then creates the chart using ggplot2
and ggthemes packages.
}
\examples{
\dontrun{
# Create a likert-style traffic light chart with percentages for sample data
sample_data <- c(1, 2, 2, 3, 3, 3, 4, 4, 4, 4, 5, 5)
ampelgrafik_likert(sample_data, "Category")
}
}
