% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_functions.R
\name{calculate_mean_sd}
\alias{calculate_mean_sd}
\title{Calculate Mean and Standard Deviation}
\usage{
calculate_mean_sd(..., na.rm = FALSE)
}
\arguments{
\item{...}{Numeric vectors or data frames for which you want to calculate the mean and standard deviation.}

\item{na.rm}{Logical. If TRUE, any NA (missing) values in the input vectors will be removed before calculations.
If FALSE (default), NA values are not removed and will yield NA results.}
}
\value{
A data frame with the following columns:
  \describe{
    \item{variable}{The names of the variables.}
    \item{mean}{The calculated means for each variable.}
    \item{sd}{The calculated standard deviations for each variable.}
  }
}
\description{
This function calculates the mean and standard deviation for each vector provided as input.
It returns a data frame with the variable names, their corresponding means, and standard deviations.
}
\examples{
\dontrun{
# Example 1: Calculate mean and standard deviation for two numeric vectors
vec1 <- c(1, 2, 3, 4, 5)
vec2 <- c(6, 7, 8, 9, 10)
calculate_mean_sd(vec1, vec2)

# Example 2: Calculate mean and standard deviation for two data frames
df1 <- data.frame(A = c(1, 2, 3), B = c(4, 5, 6))
df2 <- data.frame(A = c(7, 8, 9), B = c(10, 11, 12))
calculate_mean_sd(df1, df2)
}

}
