% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame_func.R
\name{freqtable}
\alias{freqtable}
\title{Frequency Table Generator}
\usage{
freqtable(Var)
}
\arguments{
\item{Var}{The input variable for which the frequency table is to be generated.}
}
\value{
A table containing the absolute frequency, cumulative frequency,
  percentage, and cumulative percentage of the input variable.
}
\description{
This function generates a frequency table for a given variable.
}
\examples{
\dontrun{
# Generate a frequency table for a sample variable
sample_data <- c(1, 2, 1, 3, 2, 1, 2, 3, 3)
freqtable(sample_data)
}

}
