<<<<<<< Updated upstream
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_functions.R
\name{generate_descriptive_stats_rmd}
\alias{generate_descriptive_stats_rmd}
\title{Generate an RMarkdown file with descriptive statistics for all variables in a dataframe.}
\usage{
generate_descriptive_stats_rmd(data, output_file)
}
\arguments{
\item{data}{The input dataframe for which descriptive statistics need to be generated.
The dataframe should contain a mix of numeric and categorical variables.}

\item{output_file}{The name of the output RMarkdown file that will contain the generated
descriptive statistics. The file will have the .Rmd extension.}
}
\value{
This function does not return a value but writes the RMarkdown content to the
        specified output file.
}
\description{
This function scans the input dataframe and calculates specific descriptive statistics
for each variable. For numeric variables, it calculates the summary statistics
(mean, median, standard deviation, quartiles) using \code{summary()}, \code{mean()},
\code{median()}, \code{sd()}, and \code{quantile()}. For categorical variables
(factors), it calculates frequency counts using \code{table()}.
}
\examples{
\dontrun{
# Assuming you have a dataframe called "my_data" and want to create a file called "descriptive_stats.Rmd":
generate_descriptive_stats_rmd(data = my_data, output_file = "descriptive_stats.Rmd")
}

}
=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_functions.R
\name{generate_descriptive_stats_rmd}
\alias{generate_descriptive_stats_rmd}
\title{Generate an RMarkdown file with descriptive statistics for all variables in a dataframe.}
\usage{
generate_descriptive_stats_rmd(data, output_file)
}
\arguments{
\item{data}{The input dataframe for which descriptive statistics need to be generated.
The dataframe should contain a mix of numeric and categorical variables.}

\item{output_file}{The name of the output RMarkdown file that will contain the generated
descriptive statistics. The file will have the .Rmd extension.}
}
\value{
This function does not return a value but writes the RMarkdown content to the
        specified output file.
}
\description{
This function scans the input dataframe and calculates specific descriptive statistics
for each variable. For numeric variables, it calculates the summary statistics
(mean, median, standard deviation, quartiles) using \code{summary()}, \code{mean()},
\code{median()}, \code{sd()}, and \code{quantile()}. For categorical variables
(factors), it calculates frequency counts using \code{table()}.
}
\examples{
\dontrun{
# Assuming you have a dataframe called "my_data"
and want to create a file called "descriptive_stats.Rmd":
generate_descriptive_stats_rmd(data = my_data, output_file = "descriptive_stats.Rmd")
}

}
>>>>>>> Stashed changes
