% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame_func.R
\name{plot_nNA}
\alias{plot_nNA}
\title{Plot Bar Chart with NA Count}
\usage{
plot_nNA(Data, x, y)
}
\arguments{
\item{Data}{A numeric vector containing the data to be used for plotting.}

\item{x}{The name of the variable to be plotted on the x-axis.}

\item{y}{The name of the variable to be plotted on the y-axis.}
}
\value{
A bar chart representing the distribution of the data without NA values, along with a caption
  showing the number of non-NA values and NA values in the original data.
}
\description{
This function creates a bar chart from a given data vector, where x and y represent the
variables to be plotted on the x-axis and y-axis, respectively. The function omits NA values,
counts the occurrences of each value, and includes a caption showing the total number of non-NA
values and the number of NA values in the original data. The function uses ggplot2, ggthemes, dplyr,
and magrittr packages.
}
\examples{
\dontrun{
# Create a bar chart with NA count for sample data
sample_data <- c(30, 45, NA, 10, NA, 20)
plot_nNA(sample_data, "Values", "Frequency")
}
}
