% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame_func.R
\name{plotbar}
\alias{plotbar}
\title{Plot Bar Chart}
\usage{
plotbar(Data, df, x, y)
}
\arguments{
\item{Data}{A numeric vector containing the data to be used for plotting.}

\item{df}{A data frame containing the data to be used for plotting.}

\item{x}{The name of the variable to be plotted on the x-axis.}

\item{y}{The name of the variable to be plotted on the y-axis.}
}
\value{
A bar chart representing the distribution of the data.
}
\description{
This function creates a bar chart from a given data vector and a data frame, where x and y represent the
variables to be plotted on the x-axis and y-axis, respectively. The function uses ggplot2 and ggthemes packages.
}
\examples{
\dontrun{
# Create a bar chart for sample data
sample_data <- c(30, 45, 15, 10)
sample_df <- data.frame(Category = c("A", "B", "C", "D"), Values = sample_data)
plotbar(sample_data, sample_df, "Category", "Values")
}
}
