% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame_func.R
\name{plotproz}
\alias{plotproz}
\title{Plot Percentage Bar Chart}
\usage{
plotproz(Data, x, y)
}
\arguments{
\item{Data}{A numeric vector containing the data to be plotted.}

\item{x}{The name of the variable to be plotted on the x-axis.}

\item{y}{The name of the variable to be plotted on the y-axis.}
}
\value{
A percentage bar chart representing the distribution of the data.
}
\description{
This function creates a percentage bar chart from a given data vector, where x represents the values
and y represents the variable being plotted. The function uses ggplot2 and ggthemes packages.
}
\examples{
\dontrun{
# Create a percentage bar chart for sample data
sample_data <- c(30, 45, 15, 10)
plotproz(sample_data, "Category", "Percentage")
}
}
