% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_functions.R
\name{process_file_and_save_info}
\alias{process_file_and_save_info}
\title{Process File and Save File Name Info}
\usage{
process_file_and_save_info(file_path, output_folder = "output_folder")
}
\arguments{
\item{file_path}{The path to the file to be imported and processed.}

\item{output_folder}{The path to the output folder where the file name text
file will be saved. The default is "output_folder".}
}
\value{
None (invisible NULL)
}
\description{
This function imports a file, prints its name, and saves the file name to a
text file in the specified output folder.
}
\examples{
\dontrun{
# Example usage:
file_path <- "path/to/your/file.csv"
process_file_and_save_info(file_path)
}

}
