% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame_func.R
\name{prozent}
\alias{prozent}
\title{Categorize Values into Percentage Ranges}
\usage{
prozent(df)
}
\arguments{
\item{df}{A numeric vector or data frame containing the values to be categorized.}
}
\value{
A data frame with an additional column "categories" that contains the corresponding percentage range for each value.
}
\description{
This function takes a numeric vector and categorizes its values into percentage ranges.
}
\details{
The function uses the `cut()` function to categorize the values into ten percentage ranges: 0-10, 11-20, 21-30, 31-40, 41-50, 51-60, 61-70, 71-80, 81-90, and 91-100.
}
\examples{
\dontrun{
df <- c(10, 35, 50, 75, 90)
categorized_df <- prozent(df)
}
}
