% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame_func.R
\name{prozent5}
\alias{prozent5}
\title{Categorize Values into Percentage Ranges}
\usage{
prozent5(df)
}
\arguments{
\item{df}{A numeric vector or data frame containing the values to be categorized.}
}
\value{
A data frame with an additional column "categories" that contains the corresponding percentage range for each value.
}
\description{
This function takes a numeric vector and categorizes its values into percentage ranges.
}
\details{
The function uses the `cut()` function to categorize the values into five percentage ranges: 0-20, 21-40, 41-60, 61-80, and 81-100.
}
\examples{
\dontrun{
df <- c(10, 35, 50, 75, 90)
categorized_df <- prozent5(df)
}
}
