% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame_func.R
\name{Renderengine}
\alias{Renderengine}
\title{Render R Markdown Files in Parallel}
\usage{
Renderengine(cores, PATH)
}
\arguments{
\item{cores}{Number of CPU cores to be used for parallel rendering.}

\item{PATH}{The path to the directory containing the R Markdown files to be rendered.}
}
\description{
This function renders all R Markdown files in the specified \code{PATH} using
parallel processing with the given number of \code{cores}.
}
\details{
This function uses parallel processing to render R Markdown files in
the specified \code{PATH}. Each R Markdown file will be rendered using a
separate core, which can significantly speed up the rendering process if
multiple cores are available.
}
\examples{
# Render R Markdown files in the "documents" directory using 4 cores
# Renderengine(4, "documents")

}
\seealso{
\code{\link{makeCluster}}, \code{\link{parLapply}}
}
\keyword{Markdown}
\keyword{R}
\keyword{parallel}
\keyword{rendering}
