% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_functions.R
\name{reverse_likert_polarity}
\alias{reverse_likert_polarity}
\title{Reverse the polarity of Likert scale items}
\usage{
reverse_likert_polarity(x, max_score)
}
\arguments{
\item{x}{A vector of integers representing Likert scale items.}

\item{max_score}{The maximum score of the original Likert scale.}
}
\value{
A new vector with the reversed scores.
}
\description{
This function takes a vector of Likert scale items and reverses their scoring
to convert the original scale to a new scale with reversed polarity.
}
\examples{
original_scores <- c(4, 2, 3, 5, 1)
max_score <- 5
reversed_scores <- reverse_likert_polarity(original_scores, max_score)
print(reversed_scores)

}
