% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_functions.R
\name{scale_multiple_variables}
\alias{scale_multiple_variables}
\title{Scale Multiple Variables in a Data Frame}
\usage{
scale_multiple_variables(data, var_names, scale_type = "mean")
}
\arguments{
\item{data}{A data frame containing the variables to be scaled.}

\item{var_names}{A character vector specifying the names of the variables to be scaled.}

\item{scale_type}{A character string specifying the type of scaling to be applied.
Use "sum" to scale based on the sum of the differences from the mean, or "mean" to scale based on the mean and standard deviation.
Default is "mean".}
}
\value{
A data frame with the selected variables scaled.
}
\description{
This function scales multiple variables in a data frame based on either the sum or the mean of the variables.
}
\details{
The function provides two scaling methods: "sum" and "mean".
  - When scale_type is set to "sum", the variables will be scaled by subtracting the mean and dividing by the sum of the absolute differences from the mean.
  - When scale_type is set to "mean", the variables will be scaled by subtracting the mean and dividing by the standard deviation (as implemented in the base R `scale` function).
  Note: For "sum" scaling, if the sum of absolute differences from the mean is zero for a variable, that variable will be left unchanged to avoid division by zero.
}
\examples{
# Sample data frame
data <- data.frame(
  var1 = c(10, 20, 30, 40),
  var2 = c(5, 15, 25, 35),
  var3 = c(100, 200, 300, 400)
)

# Variables to be scaled
variables_to_scale <- c("var1", "var2", "var3")

# Scale the selected variables in the data frame based on the sum
scaled_data_sum <- scale_multiple_variables(data, variables_to_scale, scale_type = "sum")

# Scale the selected variables in the data frame based on the mean
scaled_data_mean <- scale_multiple_variables(data, variables_to_scale, scale_type = "mean")

}
