% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_functions.R
\name{summary_stats_apa}
\alias{summary_stats_apa}
\title{Summarize numeric data in APA format}
\usage{
summary_stats_apa(data)
}
\arguments{
\item{data}{A numeric vector.}
}
\value{
Prints the mean and standard deviation in APA compliant format.
}
\description{
This function computes the mean and standard deviation of a numeric vector and
prints the result in APA compliant format for Markdown.
}
\examples{
summary_stats_apa(c(1, 2, 3, 4, 5))

}
